refactor: fix imports due to moving partitions
apache/gravitino#8589
