+++ clients/client-python/gravitino/api/expressions/transforms/transforms.py
@@ -21,13 +21,13 @@ from gravitino.api.expressions.expression import Expression
 from gravitino.api.expressions.literals.literal import Literal
 from gravitino.api.expressions.literals.literals import Literals
 from gravitino.api.expressions.named_reference import NamedReference
-from gravitino.api.expressions.partitions.list_partition import ListPartition
-from gravitino.api.expressions.partitions.partition import Partition
-from gravitino.api.expressions.partitions.range_partition import RangePartition
 from gravitino.api.expressions.transforms.transform import (
     SingleFieldTransform,
     Transform,
 )
+from gravitino.api.rel.partitions.list_partition import ListPartition
+from gravitino.api.rel.partitions.partition import Partition
+from gravitino.api.rel.partitions.range_partition import RangePartition
 
 
 class Transforms(Transform):
